Changed the login procedure
In the process of changing the login procedure. Leave the check as it
was, it seems to be a good way to determine if the user is genuine.
However, it cannot be used to extract a 'boolean' value that we need to
extract to determine whether the user who logged in is, in fact, an
administrator or not.